// apps/web/src/api/rest/models/OrgResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrgResponse = {
    /**
     * Organization ID
     */
    id?: string;
    /**
     * Organization name
     */
    name?: string;
    /**
     * Owner user ID
     */
    owner_id?: string;
    /**
     * Creation timestamp
     */
    created_at?: string;
};