// apps/web/src/api/rest/models/RecordResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RecordResponse = {
    id?: string;
    type_id?: string;
    org_id?: string | null;
    user_id?: string | null;
    created_by?: string;
    data?: Record<string, any>;
    created_at?: string;
    updated_at?: string;
};